requirements.txt

Flask==2.2.2
numpy>==1.22.4
tensorflow-cpu==2.10.0
werkzeug==2.2.2
gunicorn==19.9.0
jinja2==2.10.1

